// src/types/employee-values.tsx
export type EmployeeValues = {
  firstName: string;
  lastName: string;
  phone: string;
  email: string;
  employeeCode: string;
  deptId: string;
  designation: string;
  manager: string;
  hr: string;
  workingHours: string;
  noOfHolidays: string;
  isManager: boolean;
  joiningDate: string;
  grade: string;
  dob: string;
  //   team: string;
};